Remove the unused dependency on composer/package-versions-deprecated
The ProxyManagerBridge depends on friendsofphp/proxy-manager-lts which
has an optional dependency on the PackageVersions class to implement the
`ProxyManager\Version::getVersion` method on composer 1 (with a fallback
to a less precise version number).
However, the bridge has stopped using that API in favor of feature
detection, so the dependency is unused.